requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
